% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\docType{package}
\name{daptadap-package}
\alias{daptadap}
\alias{daptadap-package}
\title{daptadap}
\description{
Debug Adapter Protocol helper internals for R console workflows.
}
\author{
\strong{Maintainer}: Ilya Zarubin \email{zarubin@wiso.uni-koeln.de}

}
\keyword{internal}
